requirements.txt
sounddevice
numpy
resemblyzer
pyyaml
scipy
